requirements.txt
jsonschema>=4.22
PyYAML>=6.0
openai==2.35.1
anthropic==0.100.0
google-genai==1.75.0
